+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 **Removed features**
 
 * Will no longer update the GeoLite2 database
+* Removed subscription country determined from IP address
 
 **New features**